Blog app: add authorization rules

+++ app/controllers/posts_controller.rb
@@ -1,4 +1,6 @@
 class PostsController < ApplicationController
+  before_action :authenticate_user!
+  load_and_authorize_resource
   def index
     @user = User.find(params[:user_id])
     @user_posts = @user.posts.all.includes(:comments)
@@ -12,72 +14,51 @@ class PostsController < ApplicationController
   end
 
   def new
-    new_post = Post.new
-    respond_to do |format|
-      format.html { render :new, locals: { post: new_post } }
-    end
+    @post = Post.new
+    render :new, locals: { post: @post }
   end
 
   def create
-    new_post = Post.new(params.require(:post).permit(:title, :text))
-    new_post.author_id = current_user.id
-    new_post.save
+    @post = current_user.posts.new(post_params)
+    @post.author_id = current_user.id
 
     respond_to do |format|
       format.html do
-        if new_post.save
+        if @post.save
           # Success message
           flash.now[:success] = 'Post created successfully'
           redirect_to user_posts_path(current_user.id)
         else
           # Error message
           flash.now[:error] = 'Error: Post creation failed'
-          render :new, locals: { post: new_post }
+          render :new, locals: { post: @post }
         end
       end
     end
   end
 
-  def new_comment
-    comment = Comment.new
-    respond_to do |format|
-      format.html { render :new_comment, locals: { comment: comment } }
+  def destroy
+    @post = Post.find_by(id: params[:id], author_id: current_user.id)
+    like = Like.find_by(post_id: @post.id)
+    comment = Comment.find_by(post_id: @post.id)
+
+    like.destroy if like.present?
+    comment.destroy if comment.present?
+
+    if @post.destroy
+      # Success message
+      flash.now[:success] = 'Post deleted successfully'
+      redirect_to user_posts_path(current_user.id)
+    else
+      # Error message
+      flash.now[:error] = 'Error: Post deletion failed'
+      render :new, locals: { post: @post }
     end
   end
 
-  def create_comment
-    comment_text = params.require(:comment).permit(:text)[:text]
-    post = Post.where(author_id: current_user.id).where(id: params[:id])[0]
-    user = User.find(current_user.id)
-    new_comment = Comment.new(post: post, author: user, text: comment_text)
+  private
 
-    new_comment.save
-    puts new_comment.errors.full_messages
-    if new_comment.save
-      new_comment.update_comments_counter
-      flash[:success] = 'Success: Comment created successfully'
-      redirect_to user_post_path(user_id: current_user.id, id: post.id)
-    else
-      flash.now[:error] = 'Error: Comment creation failed'
-      render :show, locals: { post: post }
-    end
-  end
-
-  def like
-    post = Post.all.find(params[:id])
-    user = current_user
-    if Like.where(post: post, author_id: user.id).count.zero?
-      flash.now[:success] = 'Success: Post liked successfully'
-      like = Like.create(post: post, author_id: user.id)
-      like.update_likes_counter
-
-    else
-      flash.now[:success] = 'Success: you disliked this post'
-      like = Like.where(post: post, author_id: user.id)[0]
-      like.decreament_likes_counter
-      like.destroy
-    end
-
-    redirect_to user_posts_path(user_id: current_user.id)
+  def post_params
+    params.require(:post).permit(:title, :text)
   end
 end
